add more logs to git & mv generateSSHKey to test
